# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/events": {
      "get": {
        "description": "Retrieve a list of all events. Supports filtering by status and managerId. Accessible by eventmanager, client, osc, attendee, superuser.",
        "operationId": "EventsController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by status. Omit or leave empty to return all events.",
            "schema": {
              "enum": [
                "pending",
                "upcoming",
                "live",
                "completed"
              ],
              "type": "string"
            }
          },
          {
            "name": "managerId",
            "required": false,
            "in": "query",
            "description": "Filter by manager user ID (e.g. u1, u2). Pass a user ID, not a role name.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "User role (superuser | eventmanager | client | osc | attendee)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of events returned successfully"
          },
          "401": {
            "description": "Missing or invalid role header"
          },
          "403": {
            "description": "Insufficient permissions"
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Get all events",
        "tags": [
          "Events"
        ]
      },
      "post": {
        "description": "Accessible by eventmanager and client. Superuser can also create events.",
        "operationId": "EventsController_create",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Must be: superuser | eventmanager | client",
            "required": true,
            "schema": {
              "type": "string",
              "example": "eventmanager"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created successfully"
          },
          "400": {
            "description": "Validation failed"
          },
          "403": {
            "description": "Only eventmanager or client can create events"
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Create a new event",
        "tags": [
          "Events"
        ]
      }
    },
    "/events/stats/dashboard": {
      "get": {
        "description": "Returns aggregate stats: total events, RSVPs, ratings, engagement scores.",
        "operationId": "EventsController_getDashboardStats",
        "parameters": [
          {
            "name": "managerId",
            "required": false,
            "in": "query",
            "description": "Filter stats for a specific manager",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "User role (superuser | eventmanager | client | osc)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard stats returned"
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Get dashboard statistics",
        "tags": [
          "Events"
        ]
      }
    },
    "/events/{id}": {
      "get": {
        "operationId": "EventsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Event ID (e.g. e1)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "User role (superuser | eventmanager | client | osc | attendee)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event returned"
          },
          "404": {
            "description": "Event not found"
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Get event by ID",
        "tags": [
          "Events"
        ]
      },
      "put": {
        "description": "Full update of an event. Eventmanager and client only.",
        "operationId": "EventsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Event ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Must be: superuser | eventmanager | client",
            "required": true,
            "schema": {
              "type": "string",
              "example": "eventmanager"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event updated successfully"
          },
          "404": {
            "description": "Event not found"
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Update event (full replace)",
        "tags": [
          "Events"
        ]
      },
      "patch": {
        "description": "Update specific fields. Eventmanager and client only. Use for status changes (e.g. pending → live).",
        "operationId": "EventsController_patch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Event ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Must be: superuser | eventmanager | client",
            "required": true,
            "schema": {
              "type": "string",
              "example": "eventmanager"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event patched"
          },
          "404": {
            "description": "Event not found"
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Partially update event",
        "tags": [
          "Events"
        ]
      },
      "delete": {
        "description": "Permanently remove an event. Eventmanager only (or superuser).",
        "operationId": "EventsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Event ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Must be: superuser | eventmanager",
            "required": true,
            "schema": {
              "type": "string",
              "example": "eventmanager"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event deleted"
          },
          "404": {
            "description": "Event not found"
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Delete event",
        "tags": [
          "Events"
        ]
      }
    },
...
  },
  "info": {
    "title": "WEvents API",
...
    "version": "2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "role-header": {
        "type": "apiKey",
        "in": "header",
        "name": "role",
        "description": "User role (superuser | eventmanager | client | osc | attendee)"
      }
    },
    "schemas": {
      "TimelineSessionDto": {
        "type": "object",
        "properties": {
          "time": {
            "type": "string",
            "example": "09:00 AM",
            "description": "Start time of the session"
          },
          "title": {
            "type": "string",
            "example": "Keynote: Future of AI",
            "description": "Session title"
          },
          "speaker": {
            "type": "string",
            "example": "Dr. Ananya Gupta",
            "description": "Speaker name"
          },
          "location": {
            "type": "string",
            "example": "Main Auditorium",
            "description": "Session location"
          },
          "duration": {
            "type": "string",
            "example": "1 hour",
            "description": "Session duration"
          },
          "type": {
            "type": "string",
            "example": "keynote",
            "description": "Session type (keynote, panel, workshop, talk, demo, break)"
          }
        },
        "required": [
          "time",
          "title",
          "speaker"
        ]
      },
      "EventStatsDto": {
        "type": "object",
        "properties": {
          "rsvps": {
            "type": "number",
            "example": 342
          },
          "checkins": {
            "type": "number",
            "example": 298
          },
          "engagement": {
            "type": "number",
            "example": 87.5
          },
          "rating": {
            "type": "number",
            "example": 4.8
          },
          "participationTrend": {
            "example": [
              10,
              25,
              45,
              30,
              55,
              78
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "feedbackScore": {
            "type": "number",
            "example": 92
          }
        }
      },
      "CreateEventDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Global Tech Summit 2026",
            "description": "Event title"
          },
          "status": {
            "type": "string",
            "example": "upcoming",
            "enum": [
              "pending",
              "upcoming",
              "live",
              "completed"
            ],
            "description": "Event status"
          },
          "date": {
            "type": "string",
            "example": "2026-04-15",
            "description": "Event date (YYYY-MM-DD)"
          },
          "location": {
            "type": "string",
            "example": "India Habitat Centre, New Delhi",
            "description": "Event location"
          },
          "capacity": {
            "type": "number",
            "example": 500,
            "description": "Maximum capacity of the event"
          },
          "domain": {
            "type": "string",
            "example": "Client Hosted Events",
            "description": "Event domain/category"
          },
          "description": {
            "type": "string",
            "example": "Explore the future of technology with industry leaders.",
            "description": "Event description"
          },
          "managerId": {
            "type": "string",
            "example": "u1",
            "description": "ID of the event manager (superuser)"
          },
          "managerName": {
            "type": "string",
            "example": "Rajesh Kumar",
            "description": "Name of the event manager"
          },
          "timeline": {
            "description": "Event sessions/timeline",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimelineSessionDto"
            }
          },
          "stats": {
            "$ref": "#/components/schemas/EventStatsDto"
          }
        },
        "required": [
          "title",
          "status",
          "date",
          "location",
          "capacity",
          "domain",
          "description",
          "managerId",
          "managerName"
        ]
      },
      "UpdateEventDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Updated Event Title"
          },
          "status": {
            "type": "string",
            "example": "live",
            "enum": [
              "pending",
              "upcoming",
              "live",
              "completed"
            ]
          },
          "date": {
            "type": "string",
            "example": "2026-05-01"
          },
          "location": {
            "type": "string",
            "example": "Taj Hotel, Mumbai"
          },
          "capacity": {
            "type": "number",
            "example": 1000
          },
          "domain": {
            "type": "string",
            "example": "Tech Conferences"
          },
          "description": {
            "type": "string",
            "example": "Updated description."
          },
          "timeline": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimelineSessionDto"
            }
          },
          "stats": {
            "$ref": "#/components/schemas/EventStatsDto"
          }
        }
      },
...
    }
  }
}